One flip game is an android studio program that's basically a game mobile application which involves a grid of flip cards.
One flip game is an android studio program that's basically a game mobile application which involves a grid of flip cards.
It is a classic card matching animated game in Flutter.
It is designed as Flip Card memory Game in Flutter using flip card package.
This game application is targeted towards users of all ages who whish to improve their memory. It tests visual recall in an entertaining way. Users will have the option to select from a minimum of three difficulty levels. Based on this level, they will be presented a number of cards in grid form (its size increases with the difficulty mode).
The user then have to select the cards to uncover them whilst simultaneously trying to match them. The game will track either the number of tries of the user or the time that has passed in order to calculate a score and assess performance. If user reaches a new high score, then the application will display a congratulations message. In this game we also provided the option to login via external authentication.

Co-Authored-By: Amal <106037356+amalm789@users.noreply.github.com>